pc - a suggestion for dealing with the equivalent mutation
